% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf_analysis.R
\name{calculate_maf}
\alias{calculate_maf}
\title{Calculate Minor Allele Frequency (MAF)}
\usage{
calculate_maf(X)
}
\arguments{
\item{X}{A numeric matrix of genotypes (n x m), where genotypes are coded as 0, 1, or 2.}
}
\value{
A numeric vector of MAF values (1 x m).
}
\description{
This function calculates the Minor Allele Frequency (MAF) for each SNP.
}
\examples{
X <- matrix(sample(0:2, 100 * 1000, replace = TRUE), nrow = 100)  # Simulate genotype data
maf <- calculate_maf(X)
print(maf)
}
