% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_GLM.R
\name{GWAS_GLM}
\alias{GWAS_GLM}
\title{Perform GWAS using GLM}
\usage{
GWAS_GLM(y, X, C, family = gaussian(), num_cores = 1)
}
\arguments{
\item{y}{A numeric vector of phenotypes (n x 1).}

\item{X}{A numeric matrix of genotypes (n x m).}

\item{C}{A numeric matrix of covariates (n x t).}

\item{family}{A family function for the GLM (default is gaussian()).}

\item{num_cores}{Number of cores for parallel processing (default is 1).}
}
\value{
A numeric vector of p-values (1 x m) for each marker.
}
\description{
This function performs genome-wide association studies (GWAS) using a generalized linear model (GLM).
}
\examples{
y <- rnorm(100)  # Simulate phenotype data
X <- matrix(sample(0:2, 100 * 1000, replace = TRUE), nrow = 100)  # Simulate genotype data
C <- matrix(rnorm(100 * 2), nrow = 100)  # Simulate covariate data
p_values <- GWAS_GLM(y, X, C)  # Run the function
}
