% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_Cofactors.R
\name{PCA_Cofactors}
\alias{PCA_Cofactors}
\title{Perform PCA and exclude dependent PCs}
\usage{
PCA_Cofactors(X, C, threshold = 1e-06)
}
\arguments{
\item{X}{A numeric matrix of genotypes (n x m).}

\item{C}{A numeric matrix of covariates (n x t).}

\item{threshold}{A numeric value specifying the tolerance for linear dependence (default: 1e-6).}
}
\value{
A numeric matrix of selected PCs (n x k), where k is the number of non-dependent PCs.
}
\description{
This function performs principal component analysis (PCA) on genotype data and excludes PCs that are linearly dependent on covariates.
}
\examples{
X <- matrix(sample(0:2, 100 * 1000, replace = TRUE), nrow = 100)  # Simulate genotype data
C <- matrix(rnorm(100 * 2), nrow = 100)  # Simulate covariate data
pcs <- PCA_Cofactors(X, C)  # Run the function
}
