% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_and_compare.R
\name{simulate_and_compare}
\alias{simulate_and_compare}
\title{Simulate Data and Compare LDWAS and GWASbyCor}
\usage{
simulate_and_compare(
  n = 100,
  m = 1000,
  num_causal = 10,
  effect_size = 0.5,
  num_covariates = 2,
  num_replicates = 30
)
}
\arguments{
\item{n}{Number of individuals.}

\item{m}{Number of SNPs.}

\item{num_causal}{Number of causal SNPs.}

\item{effect_size}{Effect size of causal SNPs.}

\item{num_covariates}{Number of covariates.}

\item{num_replicates}{Number of replicates.}
}
\value{
A data frame containing the results for each replicate.
}
\description{
This function simulates data, runs LDWAS and GWASbyCor, and compares their performance.
}
